Feature/sc 135100/don t delete the file in case doesn t match and

+++ src/lib/download-manager.js
@@ -71,13 +71,42 @@ class DownloadManager {
 				await this._validateChecksum(filePath, expectedChecksum);
 			}
 		} catch (error) {
-			// Remove the invalid file
-			await fs.remove(filePath);
-			this.ui.write(`Removed invalid downloaded file: ${outputFileName}`);
-			throw error;
+			await this._handleInvalidChecksum({
+				error,
+				filePath,
+				displayName: outputFileName,
+				retryCallback: () => this.download({
+					url,
+					outputFileName,
+					expectedChecksum,
+					options
+				})
+			});
+		}
+		return filePath;
+	}
+
+	async _handleInvalidChecksum({ error, filePath, displayName, retryCallback }) {
+		this.ui.write(`${os.EOL}`); // Optional: visual break in terminal
+		this.ui.write(`Invalid checksum for ${displayName}`);
+
+		if (this.ui.isInteractive) {
+			const { removeFile } = await this.ui.prompt({
+				type: 'confirm',
+				name: 'removeFile',
+				message: 'Remove and retry?',
+				default: true
+			});
+
+			if (removeFile) {
+				await fs.remove(filePath);
+				this.ui.write(`Removed invalid downloaded file: ${displayName}`);
+				return retryCallback?.();
+			}
 		}
 
-		return filePath;
+		this.ui.write(`Make sure to manually delete "${filePath}" before trying again`);
+		throw error;
 	}
 
 
@@ -237,9 +266,11 @@ class DownloadManager {
 					await this._validateChecksum(cachedFilePath, expectedChecksum);
 					return cachedFilePath;
 				} catch (error) {
-					this.ui.write(`Cached file checksum mismatch for ${fileName}`);
-					await fs.remove(cachedFilePath); // Remove the invalid cached file
-					this.ui.write(`Removed invalid cached file: ${fileName}`);
+					await this._handleInvalidChecksum({
+						error: error,
+						filePath: cachedFilePath,
+						displayName: fileName
+					});
 				}
 			}
 		}
